python: move code copied from Tenjin to its own file
